Fix DORADE header parsing failure on 64-bit by using int32_t for on-disk struct fields
All DORADE on-disk format structs used `long` (8 bytes on LP64) for fields
that are 4-byte integers in the file format. This caused struct layout
mismatches (wrong offsets, padding) that made every sweep file appear as
garbage on 64-bit macOS. Changed `long` to `int32_t` across 21 header files.

+++ include/CellSpacing.h
@@ -34,6 +34,8 @@
 #ifndef INCCellSpacingh
 #define INCCellSpacingh
 
+#include <stdint.h>
+
 #ifdef OK_RPC
 
 #if defined(UNIX) && defined(sun)
@@ -49,7 +51,7 @@
 struct cell_spacing_d {
     char   cell_spacing_des[4]; /* Identifier for a cell spacing descriptor */
 				/* (ascii characters CSPD). */
-    long   cell_spacing_des_len; /* Cell Spacing descriptor length in bytes. */
+    int32_t   cell_spacing_des_len; /* Cell Spacing descriptor length in bytes. */
     short  num_segments;	/* Number of segments that contain cells of */
 				/* equal widths. */
     short  distToFirst;		/* Distance to first gate in meters. */
