+++ include/Pdata.h
@@ -3,11 +3,13 @@
 # ifndef INCPdatah
 # define INCPdatah
 
+#include <stdint.h>
+
 struct paramdata_d {
     char pdata_desc[4];	      /* parameter data descriptor identifier: ASCII */
 			      /* characters "RDAT" stand for sweep info */
 			      /* block Descriptor. */
-    long  pdata_length;       /* parameter data descriptor length in bytes. */
+    int32_t  pdata_length;       /* parameter data descriptor length in bytes. */
     char  pdata_name[8];      /*name of parameter*/
 }; /* End of Structure */
